+++ dkist/net/globus/tests/test_transfer.py
@@ -1,11 +1,14 @@
 import os
-import pathlib
+from pathlib import Path
+from unittest import mock
+from collections import namedtuple
 
 import pytest
-from globus_sdk import GlobusResponse
+from globus_sdk import GlobusHTTPResponse
+from globus_sdk.services.transfer.response.iterable import IterableTransferResponse
 
-from dkist.net.globus.transfer import (_get_speed, _process_task_events,
-                                       start_transfer_from_file_list)
+from dkist.net.globus.transfer import (_get_speed, _orchestrate_transfer_task,
+                                       _process_task_events, start_transfer_from_file_list)
 
 
 @pytest.fixture
@@ -17,38 +20,53 @@ def mock_endpoints(mocker):
                         side_effect=id_mock)
 
 
+def json_to_response(json):
+    response = mock.MagicMock()
+    response.json = lambda: json
+    return response
+
+
 @pytest.fixture
-def mock_task_event_list(mocker):
-    task_list = [
-        GlobusResponse({
-            'DATA_TYPE': 'event',
-            'code': 'STARTED',
-            'description': 'started',
-            'details':
-            '{\n  "type": "GridFTP Transfer", \n  "concurrency": 2, \n  "protocol": "Mode S"\n}',
-            'is_error': False,
-            'parent_task_id': None,
-            'time': '2019-05-16 10:13:26+00:00'
-        }),
-        GlobusResponse({
-            'DATA_TYPE': 'event',
-            'code': 'SUCCEEDED',
-            'description': 'succeeded',
-            'details': 'Scanned 100 file(s)',
-            'is_error': False,
-            'parent_task_id': None,
-            'time': '2019-05-16 10:13:24+00:00'
-        }),
-        GlobusResponse({
-            'DATA_TYPE': 'event',
-            'code': 'STARTED',
-            'description': 'started',
-            'details': 'Starting sync scan',
-            'is_error': False,
-            'parent_task_id': None,
-            'time': '2019-05-16 10:13:20+00:00'
-        })
-    ]
+def mock_task_event_list(mocker, transfer_client):
+    mock_response = GlobusHTTPResponse(
+        json_to_response(
+            {
+                "DATA": [
+                    {
+                        'DATA_TYPE': 'event',
+                        'code': 'STARTED',
+                        'description': 'started',
+                        'details':
+                        '{\n  "type": "GridFTP Transfer", \n  "concurrency": 2, \n  "protocol": "Mode S"\n}',
+                        'is_error': False,
+                        'parent_task_id': None,
+                        'time': '2019-05-16 10:13:26+00:00'},
+                    {
+                        'DATA_TYPE': 'event',
+                        'code': 'SUCCEEDED',
+                        'description': 'succeeded',
+                        'details': 'Scanned 100 file(s)',
+                        'is_error': False,
+                        'parent_task_id': None,
+                        'time': '2019-05-16 10:13:24+00:00'},
+                    {
+                        'DATA_TYPE': 'event',
+                        'code': 'STARTED',
+                        'description': 'started',
+                        'details': 'Starting sync scan',
+                        'is_error': False,
+                        'parent_task_id': None,
+                        'time': '2019-05-16 10:13:20+00:00'},
+                ],
+                "DATA_TYPE": "event_list",
+                "limit": 10,
+                "offset": 0,
+                "total": 3
+            }
+        ),
+        client=True  # Not a client object
+    )
+    task_list = IterableTransferResponse(mock_response)
     return mocker.patch("globus_sdk.TransferClient.task_event_list",
                         return_value=task_list)
 
@@ -58,7 +76,7 @@ def test_start_transfer(mocker, transfer_client, mock_endpoints):
                                return_value={"task_id": "task_id"})
     mocker.patch("globus_sdk.TransferClient.get_submission_id",
                  return_value={'value': "wibble"})
-    file_list = list(map(pathlib.Path, ["/a/name.fits", "/a/name2.fits"]))
+    file_list = list(map(Path, ["/a/name.fits", "/a/name2.fits"]))
     start_transfer_from_file_list("a", "b", "/", file_list)
     calls = mock_endpoints.call_args_list
     assert calls[0][0][0] == "a"
@@ -75,7 +93,7 @@ def test_start_transfer(mocker, transfer_client, mock_endpoints):
 def test_start_transfer_src_base(mocker, transfer_client, mock_endpoints):
     submit_mock = mocker.patch("globus_sdk.TransferClient.submit_transfer",
                                return_value={"task_id": "task_id"})
-    file_list = list(map(pathlib.Path, ["/a/b/name.fits", "/a/b/name2.fits"]))
+    file_list = list(map(Path, ["/a/b/name.fits", "/a/b/name2.fits"]))
     start_transfer_from_file_list("a", "b", "/", file_list, "/a")
     calls = mock_endpoints.call_args_list
     assert calls[0][0][0] == "a"
@@ -113,7 +131,7 @@ def test_process_event_list(transfer_client, mock_task_event_list):
 
 def test_process_event_list_message_only(transfer_client, mock_task_event_list):
     # Filter out the json event
-    prev_events = tuple(map(lambda x: tuple(x.data.items()),
+    prev_events = tuple(map(lambda x: tuple(x.items()),
                             mock_task_event_list.return_value))
     prev_events = set(prev_events[0:1])
 
@@ -143,3 +161,75 @@ def test_get_speed():
     assert speed is None
     speed = _get_speed({'code': "progress", "details": {"hello": "world"}})
     assert speed is None
+
+
+@pytest.fixture
+def orchestrate_mocks(mocker):
+    wtp = mocker.patch("dkist.net.globus.transfer.watch_transfer_progress",
+                       autospec=True)
+    mocker.patch("dkist.net.globus.transfer.get_local_endpoint_id",
+                 autospec=True, return_value="mysecretendpoint")
+    gtc = mocker.patch("dkist.net.globus.transfer.get_transfer_client",
+                       autospec=True)
+    mocker.patch("dkist.net.globus.transfer.get_data_center_endpoint_id",
+                 return_value="patched-datacenter-endpoint-id",
+                 autospec=True)
+    start = mocker.patch("dkist.net.globus.transfer.start_transfer_from_file_list",
+                         autospec=True, return_value="1234")
+    mocks = namedtuple("mocks", "watch_transfer_progress get_transfer_client start_transfer_from_file_list")
+    return mocks(wtp, gtc, start)
+
+
+@pytest.fixture
+def tfr_file_list():
+    return ["file1.fits", "file2.fits"]
+
+
+def test_ochestrate_transfer(tfr_file_list, orchestrate_mocks):
+    _orchestrate_transfer_task(tfr_file_list, recursive=False,
+                               destination_path=Path("/~/"))
+
+    orchestrate_mocks.start_transfer_from_file_list.assert_called_once_with(
+        "patched-datacenter-endpoint-id",
+        "mysecretendpoint",
+        Path("/~/"),
+        tfr_file_list,
+        recursive=False,
+    )
+
+
+def test_orchestrate_transfer_no_progress(tfr_file_list, mocker, orchestrate_mocks):
+    _orchestrate_transfer_task(tfr_file_list, recursive=False,
+                               destination_path=Path("/~/"),
+                               progress=False)
+
+    orchestrate_mocks.start_transfer_from_file_list.assert_called_once_with(
+        "patched-datacenter-endpoint-id",
+        "mysecretendpoint",
+        Path("/~/"),
+        tfr_file_list,
+        recursive=False,
+    )
+
+    orchestrate_mocks.watch_transfer_progress.assert_not_called()
+    orchestrate_mocks.get_transfer_client.return_value.task_wait.assert_called_once_with(
+        "1234",
+        timeout=1e6
+    )
+
+
+def test_orchestrate_transfer_no_wait(tfr_file_list, mocker, orchestrate_mocks):
+    _orchestrate_transfer_task(tfr_file_list, recursive=False,
+                               destination_path=Path("/~/"),
+                               wait=False)
+
+    orchestrate_mocks.start_transfer_from_file_list.assert_called_once_with(
+        "patched-datacenter-endpoint-id",
+        "mysecretendpoint",
+        Path("/~/"),
+        tfr_file_list,
+        recursive=False,
+    )
+
+    orchestrate_mocks.watch_transfer_progress.assert_not_called()
+    orchestrate_mocks.get_transfer_client.return_value.task_wait.assert_not_called()