+++ dkist/net/globus/transfer.py
@@ -6,18 +6,21 @@ import json
 import time
 import pathlib
 import datetime
+from os import PathLike
+from typing import List, Union, Literal
 
 import globus_sdk
 from parfive.utils import in_notebook
 from tqdm import tqdm, tqdm_notebook
 
-from .endpoints import auto_activate_endpoint, get_endpoint_id, get_transfer_client
+from .endpoints import (auto_activate_endpoint, get_data_center_endpoint_id,
+                        get_endpoint_id, get_local_endpoint_id, get_transfer_client)
 
 __all__ = ['watch_transfer_progress', 'start_transfer_from_file_list']
 
 
 def start_transfer_from_file_list(src_endpoint, dst_endpoint, dst_base_path, file_list,
-                                  src_base_path=None):
+                                  src_base_path=None, recursive=False):
     """
     Start a new transfer task for a list of files.
 
@@ -45,11 +48,24 @@ def start_transfer_from_file_list(src_endpoint, dst_endpoint, dst_base_path, fil
         ``eggs/`` folder will be copied to ``dst_base_path``. By default only
         the filenames are kept, and none of the directories.
 
+    recursive : `bool` or `list` of `bool`, optional
+       Controls if the path in ``file_list`` is added to the Globus task with
+       the recursive flag or not.
+       This should be `True` if the element of ``file_list`` is a directory.
+       If you need to set this per-item in ``file_list`` it should be a `list`
+       of `bool` of equal length as ``file_list``.
+
     Returns
     -------
     `str`
         Task ID.
     """
+    if isinstance(recursive, bool):
+        recursive = [recursive] * len(file_list)
+    if len(recursive) != len(file_list):
+        raise ValueError(
+            "The length of recursive should equal the length of file_list when specified as a list."
+        )
 
     # Get a transfer client instance
     tc = get_transfer_client()
@@ -79,8 +95,8 @@ def start_transfer_from_file_list(src_endpoint, dst_endpoint, dst_base_path, fil
             src_filepath = src_file.relative_to(src_base_path)
         dst_file_list.append(dst_base_path / src_filepath)
 
-    for src_file, dst_file in zip(src_file_list, dst_file_list):
-        transfer_manifest.add_item(str(src_file), str(dst_file))
+    for src_file, dst_file, rec in zip(src_file_list, dst_file_list, recursive):
+        transfer_manifest.add_item(str(src_file), str(dst_file), recursive=rec)
 
     return tc.submit_transfer(transfer_manifest)["task_id"]
 
@@ -114,8 +130,8 @@ def _process_task_events(task_id, prev_events, tfr_client):
     """
 
     # Convert all the events into a (key, value) tuple pair
-    events = set(map(lambda x: tuple(x.data.items()),
-                     tfr_client.task_event_list(task_id, None)))
+    events = set(map(lambda x: tuple(x.items()),
+                     tfr_client.task_event_list(task_id)))
     # Drop all events we have seen before
     new_events = events.difference(prev_events)
 
@@ -186,37 +202,131 @@ def watch_transfer_progress(task_id, tfr_client, poll_interval=5,
     progress = get_progress_bar(unit="file",
                                 dynamic_ncols=True,
                                 total=initial_n)
-    while True:
-        (prev_events,
-         json_events,
-         message_events) = _process_task_events(task_id, prev_events, tfr_client)
 
-        if ('code', 'STARTED') not in prev_events and not started:
-            started = True
-            progress.write("PENDING: Starting Transfer")
+    try:
+        while True:
+            (prev_events,
+             json_events,
+             message_events) = _process_task_events(task_id, prev_events, tfr_client)
 
-        # Print status messages if verbose or if they are errors
-        for event in message_events:
-            if event['is_error'] or verbose:
-                progress.write(f"{event['code']}: {event['details']}")
+            if ('code', 'STARTED') not in prev_events and not started:
+                started = True
+                progress.write("PENDING: Starting Transfer")
 
-        # Extract and calculate the transfer speed from the event list
-        speed = (list(map(_get_speed, json_events)) or [None])[0]
-        speed = f"{speed} Mb/s" if speed else ""
-        if speed:
-            progress.set_postfix_str(speed)
+            # Print status messages if verbose or if they are errors
+            for event in message_events:
+                if event['is_error'] or verbose:
+                    progress.write(f"{event['code']}: {event['details']}")
 
-        # Get the status of the task to see how many files we have processed.
-        task = tfr_client.get_task(task_id)
-        status = task['status']
-        progress.total = task['files']
-        progress.update((task['files_skipped'] + task['files_transferred']) - progress.n)
+            for event in json_events:
+                # This block was coded off one example, as I can't find any
+                # documentation of the structure of events. This is why it's
+                # all very tolerant of missing keys etc.
+                if event["is_error"] or verbose:
+                    progress.write(f"{event.get('code', '')}: {event.get('description', '')}")
+                    details = event.get("details", {})
+                    if "error" in details and "body" in details["error"]:
+                        extra_message = details["error"]["body"]
+                        if "context" in details:
+                            context = details["context"][0]
+                            operation = context.get("operation", "")
+                            path = context.get("path", "")
+                            extra_message = f"{operation} | {path}".strip() + " | " + extra_message
+                        progress.write(extra_message)
 
-        # If the status of the task is not active we are finished.
-        if status != "ACTIVE":
-            progress.write(f"Task completed with {status} status.")
-            progress.close()
-            break
+            # Extract and calculate the transfer speed from the event list
+            speed = (list(map(_get_speed, json_events)) or [None])[0]
+            speed = f"{speed} Mb/s" if speed else ""
+            if speed:
+                progress.set_postfix_str(speed)
 
-        # Wait for next poll
-        time.sleep(poll_interval)
+            # Get the status of the task to see how many files we have processed.
+            task = tfr_client.get_task(task_id)
+            status = task['status']
+            progress.total = task['files']
+            progress.update((task['files_skipped'] + task['files_transferred']) - progress.n)
+
+            # If the status of the task is not active we are finished.
+            if status != "ACTIVE":
+                progress.write(f"Task completed with {status} status.")
+                progress.close()
+                break
+
+            # Wait for next poll
+            time.sleep(poll_interval)
+
+    except KeyboardInterrupt:
+        progress.write("Cancelling Task")
+        task = tfr_client.cancel_task(task_id)
+        progress.close()
+
+
+def _orchestrate_transfer_task(file_list: List[PathLike],
+                               recursive: List[bool],
+                               destination_path: PathLike = "/~/",
+                               destination_endpoint: str = None,
+                               *,
+                               progress: Union[bool, Literal["verbose"]] = True,
+                               wait: bool = True):
+    """
+    Transfer the files given in file_list to the path on ``destination_endpoint``.
+
+    Parameters
+    ----------
+    file_list
+        The list of file paths on the source endpoint to transfer to the
+        ``destination_endpoint``.
+
+    recursive
+       Controls if the path in ``file_list`` is added to the Globus task with
+       the recursive flag or not.
+
+    destination_path
+        The path to the directory containing the dataset on the destination
+        endpoint.
+
+    destination_endpoint
+        A unique specifier for a Globus endpoint. If `None` a local
+        endpoint will be used if it can be found, otherwise an error will
+        be raised. See `~dkist.net.globus.get_endpoint_id` for valid
+        endpoint specifiers.
+
+    progress
+        If `True` status information and a progress bar will be displayed
+        while waiting for the transfer to complete.
+        If ``progress="verbose"`` then all globus events generated during the
+        transfer will be shown (by default only error messages are shown.)
+        This keyword is only used if ``wait=True``.
+
+    wait
+       If `False` then the function will return while the Globus transfer task
+       is still running. Setting ``wait=False`` implies ``progress=False``.
+
+    Returns
+    -------
+    destination_path
+        The path to the directory containing the dataset on the destination
+        endpoint.
+
+    """
+
+    if not destination_endpoint:
+        destination_endpoint = get_local_endpoint_id()
+
+    task_id = start_transfer_from_file_list(get_data_center_endpoint_id(),
+                                            destination_endpoint,
+                                            destination_path,
+                                            file_list,
+                                            recursive=recursive)
+
+    tc = get_transfer_client()
+
+    if wait:
+        if progress:
+            watch_transfer_progress(task_id,
+                                    tc,
+                                    verbose=progress == "verbose")
+        else:
+            tc.task_wait(task_id, timeout=1e6)
+
+    return destination_path